+++ src/constellation/index.go
@@ -18,7 +18,7 @@ func Init() {
 
 	// if date is > 1st of January 2024
 	timeNow := time.Now()
-	if  timeNow.Year() > 2024 || (timeNow.Year() == 2024 && timeNow.Month() > 6) {
+	if  timeNow.Year() > 2024 || (timeNow.Year() == 2024 && timeNow.Month() > 9) {
 		utils.Error("Constellation: this preview version has expired, please update to use the lastest version of Constellation.", nil)
 		// disable constellation
 		configFile := utils.ReadConfigFromFile()